corrected a case sensitive issue with cmd/ttapp module
removed the load config function from the api and adapted the functional test
