+++ api/config.go
@@ -1,29 +1,6 @@
 package api
 
-import (
-	"encoding/json"
-	"fmt"
-	"os"
-)
-
 type Config struct {
 	ServerAddress string `json:"server_address"`
 	ServerPort    string `json:"server_port"`
-}
-
-func LoadConfig(filename string) (*Config, error) {
-	file, err := os.Open(filename)
-	if err != nil {
-		return nil, fmt.Errorf("could not open config file: %w", err)
-	}
-	defer file.Close()
-
-	var config Config
-	decoder := json.NewDecoder(file)
-	err = decoder.Decode(&config)
-	if err != nil {
-		return nil, fmt.Errorf("could not decode config JSON: %w", err)
-	}
-
-	return &config, nil
 }
